Move Check Initialization to its own file
