+++ src/main/scala/firrtl/passes/Checks.scala
@@ -742,54 +742,3 @@ object CheckWidths extends Pass with StanzaPass {
       c
    }
 }
-
-object CheckInitialization extends Pass with StanzaPass {
-   def name = "Check Initialization"
-   var mname = ""
-   class RefNotInitialized (info:Info, name:String) extends PassException(s"${info} : [module ${mname}  Reference ${name} is not fully initialized.")
-   def run (c:Circuit): Circuit = {
-      val errors = ArrayBuffer[PassException]()
-      def check_init_m (m:InModule) : Unit = {
-         def get_name (e:Expression) : String = {
-            (e) match { 
-               case (e:WRef) => e.name
-               case (e:WSubField) => get_name(e.exp) + "." + e.name
-               case (e:WSubIndex) => get_name(e.exp) + "[" + e.value + "]"
-               case (e) => error("Shouldn't be here"); ""
-            }
-         }
-         def has_voidQ (e:Expression) : Boolean = {
-            var void = false
-            def has_void (e:Expression) : Expression = {
-               (e) match { 
-                  case (e:WVoid) => void = true; e
-                  case (e) => e map (has_void)
-               }
-            }
-            has_void(e)
-            void
-         }
-         def check_init_s (s:Stmt) : Stmt = {
-            (s) match { 
-               case (s:Connect) => {
-                  if (has_voidQ(s.exp)) errors += new RefNotInitialized(s.info,get_name(s.loc))
-                  s
-               }
-               case (s) => s map (check_init_s)
-            }
-         }
-         check_init_s(m.body)
-      }
-         
-      for (m <- c.modules) {
-         mname = m.name
-         (m) match  { 
-            case (m:InModule) => check_init_m(m)
-            case (m) => false
-         }
-      }
-
-    if (errors.nonEmpty) throw new PassExceptions(errors)
-    c
-   }
-}
